Update to fix a bug, refactor code, update docstrings
There was an error when seeding the database with CSV files, which was
due to an indentation error in the load_csv function in read_files.py.

Everything now works to seed tables with no dependencies (ie: concerns,
categories, ingredients, skincare_steps, and skintypes).

Will continue to work on seeding the products table. This will also
require seeding both the ingredients and the products_ingredients
tables.