Complete early goal to seed database with sample data
Using new sample data (samp_products.csv instead of
sample_products.csv) to test and successfully fix
an issue with reading files. The product_url field
is now the last column to avoid incorrect file
parsing. (It was registering the comma between the
product_url and the product_type as part of the url
itself, resulting in index issues.)

The following tables are now seeding correctly:
  - categories
  - concerns
  - skintypes
  - skincare_steps
  - products (but will need to add category_id)
  - ingredients
  - product_ingredients