Update, refactor, and verify that proper database seeding
The ingredients table will now only load new rows of data if those
ingredients do not yet exist in the table.

Will need to add ability to check if product already exists in the
products table (fix function in crud.py).

Also added convert_price function in crud.py, but it is not in use.