Complete add_to_cabinet.js (add multiple products from search results)
I played around with a livesearch.js file to listen for text input
within the search keyword input in order to send queries for similar
product names without submitting the form. However, this is not fully
functional since ORMs are not serializable and therefore cannot be
retrieved as a response in the JS file. Possible considerations:
  - marshmallow, marshmallow-sqlalchemy, or flask-marshmallow
  - add a @property to serialize the ORM classes in model.py

Dataset has some irregularities, but I was able to find a UNII code
dataset from the FDA. I was going to load this into my db to add a
UNII code field of some kind to my ingredients table, but the dataset
is nearly 60 MB - in a txt file! Might clean up dataset with pandas,
if I end up doing this sometime soon.

New routes on server.py include:
  - '/test' - to test livesearch
  - '/livesearch' - to test AJAX portion of livesearch
  - '/add_to_cabinet' - to add products from results page to cabinet

Modified '/products/search' route on server.py to allow queries to
utilize ORDER BY and LIMIT when searching for skincare products.