Setup test version of routine, add serialize property to ORM (livesearch)
Over the weekend, spent a long time trying to setup a local dev
environment (to try to move away from the remote container on Docker,
that was possibly causing some issues). This did not work due to
version control issues for virtualenv, Python3.10, pandas, and psql.
I will now work on the remote container for the rest of the project.

For livesearch: added serialize property to the Product ORM class in
model.py. Will test shortly. May use React in the future.

For skincare routines: added new server route to test development of
individual skincare routines using jQuery. Steps are now snappable,
draggable, and sortable. Each step contains a dropdown menu, which I
will link up to the db to get the list of skincare products in the
user's cabinet that are of the same product type for that step
(eg: get a list of cleansers from the cabinet for the cleanser step).