Add db summary to homepage (via crud); setup maintenance.py
Summary info for the database is calculated in the crud file, but this
takes ~50 seconds to calculate and load the data onto the homepage.
This will be replaced with a CSV reader to read the file generated by
maintenance.py in the new database directory.
  - only run maintenance.py after new product datasets are loaded into the db
  - otherwise, db_summary.csv can be read directly

Also added functions to get the current UTC datetime and to convert to
Pacific Time using the pytz library. Will add to ORM classes later.