Save progress on saving routine via AJAX
Added a button to create AM and PM skincare routines. This allows
routines.js to generate draggable skincare steps and to populate
the dropdown menus for each step with skincare products from the
user's cabinet for each category type (all via AJAX and jQuery).

Added a button to save the routine and to make a POST request to
the server via the '/routine' route. Need to figure out how to
make the data send in a clear format that is easy to parse, so
the data can be used to create new rows in the appropriate tables
of the database.