Restructure directories and imports; minor refactoring
The db_summary.csv file is now located in the static directory
to allow for easier integration with Flask. The maintenance.py
file is updated with the new filepath.

The crud.py and model.py files are now located in the database
directory and contain the appropriate import statements.

The seed_db.py no longer contains the non-functional import
statements; those lines of code were able to add the parent
directory to sys.path, but this did not help with importation
from parent directories...