Update db seeding/maintenance; add file about datasets
A new test db (project_test_2) was seeded yesterday with the Kaggle
dataset into the PostgreSQL database called "project_test_2". The data
was loaded successfully and can be queried and displayed on the website.

This involved significant refactoring of the create_product_cascade and
create_ingredients_cascade functions in crud.py.

Added a file describing the datasets and resources reviewed, including
government agencies and individuals online. Please contact me if you
would like any of your data to be removed from this project!

Other significant changes:
  - removed several redundant functions from crud.py (lines 269-322)
    that were already moved to maintenance.py and commited on 10/25
  - modified the way connect_to_db is used
    - added _DB_NAME_ at the top of the file to make it easier to edit
      when switching between different test databases